Lookup table for building strip index from phi coordinate

+++ LinearizedTrackFit/interface/Stub.h
@@ -20,14 +20,14 @@ class Stub
   double R() const { return R_; }
   double z() const { return z_; }
   int layer() const { return layer_; }
-  int strip() const { return strip_; }
+  float strip() const { return strip_; }
 
  private:
   double phi_;
   double R_;
   double z_;
   int layer_;
-  int strip_;
+  float strip_;
 };
